+++ NEWS.md
@@ -7,3 +7,5 @@
 - Addition of georeferenced data for Brazil (at city, state and region levels) and world (at country level).
 
 - Implementation of add_geo() function to add georeferenced data to the downloaded COVID19 data.
+
+- Lubridate package removed from the Imports field in the DESCRIPTION file.